Move System.Text.Json implementation into separate csproj.
