+++ README.md
@@ -182,13 +182,16 @@ This leads to the following list of naming conventions:
 
 ## Should I use this Library?
 
-It's a library that is still evolving rapidly. Try at your own risk or - even better - contribute :-)
+The API seems to be stabilising and serialization is supported. Hence, It's worth giving it a try :smiley:
 
 ## ToDo
 
 - [x] Enable [binary prefixes](https://en.wikipedia.org/wiki/Binary_prefix).
   - Enabling things like "KiB", i.e. "kibi Byte".
-- [x] Enable serialisation
+- Enable serialisation using
+  - [x] [System.Text.Json](quantities.serialization/quantities.serialization.text.json/Extensions.cs)
+  - [ ] `Newtonsoft`
+  - [ ] Google Protocol Buffers
 - [ ] Extend unit tests
 - [ ] More rigours benchmarking
 - [ ] Add more quantities
