+++ quantities.benchmark/Convenience.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Text.Json;
+using Quantities.Serialization.Text.Json;
 
 namespace Quantities.Benchmark;
