+++ quantities.benchmark/quantities.benchmark.csproj
@@ -19,5 +19,6 @@
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\quantities\quantities.csproj" />
+    <ProjectReference Include="..\quantities.serialization\text.json\quantities.serialization.text.json.csproj" />
   </ItemGroup>
 </Project>
