+++ quantities.serialization/text.json/Deserializer.cs
@@ -1,9 +1,7 @@
 using System.Text.Json;
 using static System.Text.Json.JsonTokenType;
 
-namespace Quantities.Serialization;
-
-public readonly record struct QuantityModel(String System, String? Prefix, String Unit);
+namespace Quantities.Serialization.Text.Json;
 
 internal static class Deserializer
 {
@@ -16,7 +14,7 @@ internal static class Deserializer
 
     public static String ReadNameOf(this ref Utf8JsonReader reader, JsonTokenType tokenType)
     {
-        if (MoveNext(ref reader, tokenType)) {
+        if (reader.MoveNext(tokenType)) {
             return reader.GetString() ?? throw new InvalidDataException($"Encountered empty property name for token of type '{tokenType}'.");
         }
         throw new InvalidDataException($"Failed finding next token of type '{tokenType}'."); ;
@@ -24,7 +22,7 @@ internal static class Deserializer
 
     public static Double ReadNumber(this ref Utf8JsonReader reader)
     {
-        if (MoveNext(ref reader, Number)) {
+        if (reader.MoveNext(Number)) {
             return reader.GetDouble();
         }
         throw new InvalidDataException("Failed finding next number");
@@ -32,7 +30,7 @@ internal static class Deserializer
 
     public static String ReadString(this ref Utf8JsonReader reader)
     {
-        if (MoveNext(ref reader, JsonTokenType.String)) {
+        if (reader.MoveNext(JsonTokenType.String)) {
             return reader.GetString() ?? throw new InvalidDataException("Expected a non null value for a string.");
         }
         throw new InvalidDataException("Failed finding next string");
@@ -45,11 +43,11 @@ internal static class Deserializer
 
     public static QuantityModel Read(this ref Utf8JsonReader reader, String system)
     {
-        MoveNext(ref reader, StartObject);
+        reader.MoveNext(StartObject);
         String? unit = null, prefix = null;
         while (unit == null) {
-            var propertyName = ReadNameOf(ref reader, PropertyName);
-            var propertyValue = ReadString(ref reader);
+            var propertyName = reader.ReadNameOf(PropertyName);
+            var propertyValue = reader.ReadString();
             if (propertyName is nameof(prefix)) {
                 prefix = propertyValue;
                 continue;
