+++ quantities.test/quantities.test.csproj
@@ -20,5 +20,6 @@
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\quantities\quantities.csproj" />
+    <ProjectReference Include="..\quantities.serialization\text.json\quantities.serialization.text.json.csproj" />
   </ItemGroup>
 </Project>
